Fix: 등록, 메인, 어드민 페이지 이슈 수정

+++ front/public/js/logout.js
@@ -1,15 +1,16 @@
 const logoutBtn = document.querySelector(".logout-btn");
 
-logoutBtn.addEventListener("click", async() => {
-    
-    const logoutConfirm = confirm("로그아웃 하시겠습니까?");
-    if(!logoutConfirm) {return};
+if(logoutBtn) {
+    logoutBtn.addEventListener("click", async() => {
+        
+        const logoutConfirm = confirm("로그아웃 하시겠습니까?");
+        if(!logoutConfirm) {return};
 
-    try {
-        const { data } = await axios.delete("/auth/logout");
-        window.location.href = data.message
-    } catch (error) {
-        alert("로그아웃에 실패하였습니다.")
-    };
-    
-});
+        try {
+            const { data } = await axios.delete("/auth/logout");
+            window.location.href = data.message
+        } catch (error) {
+            alert("로그아웃에 실패하였습니다.")
+        };
+    });
+}
